Moved single process extraction engine to sub module
