+++ plaso/single_process/extraction_engine.py
